Implement collision system
This commit modifies objects to give them collision aspect. World
objects can optionaly have a collision body and a collision group, and
can react to collisions.

The world runs a collision loop after updating.

In order to test the system, the collision box has been added to player
bullets and to boss (?!) and they now react to each other.