Recycle destroyed shots to create new ones
This commit adds a pool mechanism in the world object. If object
initialization is separated between a "constructor" method that just
setup the basic generic structure of an object, and a "init" method that
truly set the parameters, then the world is able to put dead objects in
a pool, and recycle them to create new objects instead of recreating the
whole structure.

The gatling and ship bullets have been rewritten to use this pool
system. This greatly improve performances on mobile devices.